fix: combine pagination-range to Pagination.tsx
